backend/requirements.txt
# Core
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
flask-sqlalchemy>=3.1.1
flask-limiter==3.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.35.1

# AI / ML
transformers>=4.38.0
torch>=2.6.0
sentencepiece>=0.1.99
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
numpy>=1.26.4
scikit-learn>=1.4.0

# Safety / NLP
better-profanity==0.7.0
nltk==3.8.1

# Database
sqlalchemy>=2.0.36

# Audio / TTS
gtts==2.5.1
pydub==0.25.1

# Testing
pytest==7.4.0
pytest-flask==1.3.0

# Utilities
python-dateutil==2.9.0
